upgrade to react query

+++ frontend/src/screens/ProfileScreen.tsx
@@ -1,80 +1,41 @@
-import React, { useContext, useReducer, useState } from 'react'
+import React, { useContext, useState } from 'react'
 import { Helmet } from 'react-helmet-async'
 import Form from 'react-bootstrap/Form'
 import Button from 'react-bootstrap/Button'
 import { Store } from '../Store'
 import { toast } from 'react-toastify'
 import { getError } from '../utils'
-import axios from 'axios'
-import { Navigate, useNavigate } from 'react-router-dom'
 import { ApiError } from '../types/ApiError'
 import LoadingBox from '../components/LoadingBox'
-
-type State = {
-  loadingUpdate: boolean
-  error: string
-}
-type Action =
-  | { type: 'UPDATE_REQUEST' }
-  | {
-      type: 'UPDATE_SUCCESS'
-    }
-  | { type: 'UPDATE_FAIL' }
-const initialState: State = {
-  loadingUpdate: false,
-  error: '',
-}
-const reducer = (state: State, action: Action) => {
-  switch (action.type) {
-    case 'UPDATE_REQUEST':
-      return { ...state, loadingUpdate: true }
-    case 'UPDATE_SUCCESS':
-      return { ...state, loadingUpdate: false }
-    case 'UPDATE_FAIL':
-      return { ...state, loadingUpdate: false }
-
-    default:
-      return state
-  }
-}
+import { useUpdateProfileMutation } from '../hooks/userHooks'
 
 export default function ProfileScreen() {
   const { state, dispatch: ctxDispatch } = useContext(Store)
   const { userInfo } = state
-  const navigate = useNavigate()
   const [name, setName] = useState(userInfo!.name)
   const [email, setEmail] = useState(userInfo!.email)
   const [password, setPassword] = useState('')
   const [confirmPassword, setConfirmPassword] = useState('')
 
-  const [{ loadingUpdate }, dispatch] = useReducer<
-    React.Reducer<State, Action>
-  >(reducer, initialState)
+  const { mutateAsync: updateProfile, isLoading } = useUpdateProfileMutation()
 
   const submitHandler = async (e: React.SyntheticEvent) => {
     e.preventDefault()
     try {
-      const { data } = await axios.put(
-        '/api/users/profile',
-        {
-          name,
-          email,
-          password,
-        },
-        {
-          headers: { Authorization: `Bearer ${userInfo!.token}` },
-        }
-      )
-      dispatch({
-        type: 'UPDATE_SUCCESS',
+      if (password !== confirmPassword) {
+        toast.error('Passwords do not match')
+        return
+      }
+      const data = await updateProfile({
+        name,
+        email,
+        password,
       })
+
       ctxDispatch({ type: 'USER_SIGNIN', payload: data })
       localStorage.setItem('userInfo', JSON.stringify(data))
       toast.success('User updated successfully')
     } catch (err) {
-      dispatch({
-        type: 'UPDATE_FAIL',
-      })
       toast.error(getError(err as ApiError))
     }
   }
@@ -118,10 +79,10 @@ export default function ProfileScreen() {
           />
         </Form.Group>
         <div className="mb-3">
-          <Button disabled={loadingUpdate} type="submit">
+          <Button disabled={isLoading} type="submit">
             Update
           </Button>
-          {loadingUpdate && <LoadingBox></LoadingBox>}
+          {isLoading && <LoadingBox></LoadingBox>}
         </div>
       </form>
     </div>
